Controllable Generation
In this notebook, we implement a GAN controllability method using gradients from a classifier. BY training a classifier to recognize a relevant feature, we can use it to change the generator's input (z-vectors) to make it generate images with more or less of that feature. We will be using CelebA dataset.